move ensureChain to the SDK
