+++ packages/indexer/features/asset-types/search/PlaceForm.tsx
@@ -10,12 +10,6 @@ import {
   useToast,
 } from "@chakra-ui/react";
 import { DecodedAssetId, GeohashBits } from "@geospatialregistry/sdk";
-import { hexValue } from "@ethersproject/bytes";
-import {
-  ExternalProvider,
-  JsonRpcSigner,
-  Web3Provider,
-} from "@ethersproject/providers";
 
 import { gsr } from "~/features/gsr/gsr-contract";
 import { gsrIndexer } from "~/features/gsr/gsr-indexer";
@@ -66,7 +60,6 @@ export const PlaceForm: FunctionComponent<PlaceFormProps> = ({
 
     try {
       const signer = provider.getSigner();
-      await ensureChain(signer);
 
       const { sync } = await gsr.place(signer, decodedAssetId, newLocation, {
         timeRange: { start: 0, end: 0 },
@@ -138,92 +131,3 @@ export const PlaceForm: FunctionComponent<PlaceFormProps> = ({
     </Box>
   );
 };
-
-/** Ensure the user is on the correct chain. */
-const ensureChain = async (signer: JsonRpcSigner) => {
-  if (!signer) {
-    throw new Error("No signer");
-  }
-
-  const gsrChainId = gsrIndexer.chainId;
-
-  // If chain is already correct, do nothing.
-  if ((await signer.getChainId()) === gsrChainId) {
-    return;
-  }
-
-  if (gsrChainId === 1337) {
-    throw new Error(
-      "Cannot manually switch to devnet. manually switch to chain 1337."
-    );
-  }
-
-  const externalProvider: ExternalProvider = (signer.provider as any)?.provider;
-  if (!externalProvider) return;
-
-  // Get a provider that can request chain changes.
-  const ethereum = new Web3Provider(externalProvider).provider;
-  if (!ethereum?.request) return;
-
-  const hexChainId = hexValue(gsrChainId);
-
-  try {
-    await ethereum.request({
-      method: "wallet_switchEthereumChain",
-      params: [{ chainId: hexChainId }],
-    });
-  } catch (switchError: any) {
-    // This error code indicates that the chain has not been added to MetaMask.
-    if ("code" in switchError && switchError.code === 4902) {
-      try {
-        await ethereum.request({
-          method: "wallet_addEthereumChain",
-          params: [
-            {
-              chainId: hexChainId,
-              chainName: chainNames[gsrChainId],
-              rpcUrls: [chainRpcUrls[gsrChainId]],
-              nativeCurrency: {
-                name: "MATIC",
-                symbol: "MATIC",
-                decimals: 18,
-              },
-              blockExplorerUrls: [chainIdExplorer(gsrChainId)],
-            },
-          ],
-        });
-      } catch (addError) {
-        throw new Error(
-          "Please switch to the Polygon network to publish with the GSR."
-        );
-      }
-    } else {
-      throw new Error(
-        "Please switch to the Polygon network to publish with the GSR."
-      );
-    }
-  }
-};
-
-const chainNames: Record<number, string> = {
-  137: "Polygon Mainnet",
-  80001: "Mumbai Testnet",
-  1337: "Hardhat Network",
-};
-
-const chainRpcUrls: Record<number, string> = {
-  137: "https://polygon-rpc.com/",
-  80001: "https://rpc-mumbai.maticvigil.com",
-  1337: "http://127.0.0.1:8545/",
-};
-
-/** Look up block explorer by chainId */
-export function chainIdExplorer(chainId: number): string {
-  if (chainId === 80001) {
-    return "https://mumbai.polygonscan.com";
-  }
-  if (chainId === 137) {
-    return "https://polygonscan.com";
-  }
-  return "";
-}
