+++ packages/sdk/src/gsr-contract.ts
@@ -12,6 +12,7 @@ import {
   AssetTypeVerifier,
   DecodedAssetId,
 } from "./asset-types/AssetTypeVerifier";
+import { ensureActiveChain } from "./ensure-chain";
 import { GeohashBits } from "./geohash";
 import { GsrIndexer } from "./gsr-indexer";
 import {
@@ -75,10 +76,15 @@ export interface GsrContractOpts {
 
 /** Make requests to the GSR smart contract using decoded asset IDs. */
 export class GsrContract {
+  /** Reference to the contract. */
   public contract: GeoSpatialRegistry;
+  /** RPC Provider for read access to the contract.  */
   public gsrProvider: Provider;
-
+  /** Reference to the asset type verifiers. */
   public verifier: AssetTypeVerifier;
+  /** ChainId contract calls will go to. */
+  public chainId: number;
+  /** Reference to the indexer for the same Chain ID */
   private indexer: GsrIndexer;
 
   constructor(
@@ -90,10 +96,12 @@ export class GsrContract {
       indexer = new GsrIndexer(chainId),
     }: GsrContractOpts = {}
   ) {
-    this.gsrProvider =
-      customGsrProvider || getChainProvider(chainId, providerKeys);
+    this.chainId = chainId;
 
-    if (indexer.chainId !== chainId) {
+    this.gsrProvider =
+      customGsrProvider || getChainProvider(this.chainId, providerKeys);
+
+    if (indexer.chainId !== this.chainId) {
       // eslint-disable-next-line no-console
       console.warn("GsrIndexer and GsrContract have different chain IDs");
     }
@@ -101,7 +109,7 @@ export class GsrContract {
     this.indexer = indexer;
     this.verifier = new AssetTypeVerifier(providerKeys);
 
-    const address = customGsrAddress || GsrAddress[chainId];
+    const address = customGsrAddress || GsrAddress[this.chainId];
 
     this.contract = GeoSpatialRegistry__factory.connect(
       address,
@@ -243,12 +251,21 @@ export class GsrContract {
     {
       timeRange = { start: 0, end: 0 },
       sceneUri,
+      ensureChain = true,
     }: {
+      /** Optionally add a start and/or end date that the placement is valid within. */
       timeRange?: TimeRange;
+      /** Optionally pass a sceneUri to set with the placement. */
       sceneUri?: string;
-      /** If true don't resolve the promise until the placement is minted and synced.  */
+      /** If false, do not verify the chain ID of the wallet. */
+      ensureChain?: boolean;
     } = {}
   ): Promise<ContractCallResponse> {
+    //
+    if (ensureChain) {
+      await ensureActiveChain(signer, this.chainId);
+    }
+
     const encodedAssetId = this.verifier.encodeAssetId(decodedAssetId);
 
     const tx = sceneUri
